+++ src/modules/browse/explore/components/DietInfoCard.tsx
@@ -4,7 +4,7 @@ import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
 import { Button, CardActionArea, CardActions } from '@mui/material';
-import Diet from '../../general/interfaces/Diet';
+import Diet from '../../../general/interfaces/Diet';
 
 export default function MultiActionAreaCard(diet: Diet) {
   return (
