+++ src/modules/browse/explore/components/Slider/index.tsx
